api_py/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.21
psycopg[binary]==3.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
